Resolve TS compiler errors
For the "wrapper" components, the error was that the `attachments` object did not accept keys of the `string` type, but the value of the `id_ref` variable was of the `string` type. I resolved these errors by replacing the `[]` index operator with the `Object.getOwnPropertyDescriptor` function.

For the error with the top-level `await` statement in the `src/templates/projects_config.ts` file, I bumped the compiler target to `es2017`.

Finally, the remaining errors were all variables with the implicit `any` type, which I made explicit. These could be improved by typing the output of the `PouchDB.Database<T>.get` function.