Fix issues that cause "A component is changing an uncontrolled input to be controlled." warnings
The issue was caused by passing `null` or `undefined` to an `<input>` element on the first render and then passing a defined value, resulting in the component changing from uncontrolled to controlled between renders.

For checkboxes, the issue was caused by modifying the value of a `useState`-controlled variable between renders.