Refactor types
The purpose of this commit is to "push" the types from the `src/types/*.ts` files into the `src/components/*.tsx` files, refactor the code so that redundant types are removed, and then to move reusable types back into `src/types/database.types.ts` file.

The result is that the `src/components/store.tsx` file can now be type annotated and refactored.