Fix position of title when back button is visible
When the back button is visible, the flex layout is resized automatically and so the position of the title is shifted right.

This commit adds a `<div>` element with fixed dimensions so that the position of the title is shifted left.

The DOM element ID for the new `<div>` element is "settings-bottom-container" to allude to the future incorporation of a "Settings" screen.