Refactor new/edit project
This commit alters the process for creating new projects. Instead of creating a new `Project` document in PouchDB and then requesting that the user enter the new document's name, a modal dialog is displayed, a valid name is entered, and then the new `Project` document is created.

The consequence of this commit is that the `DocNameInput` and `SaveCancelButton` components can be removed.