Link to PouchDB document for "doe_workflow_combustion_appliance_safety_tests" templates
The `Select` component is refactored to accept an array of key-value pairs as the `options` prop.

The `InstallationSelectWrapper` component (exposed to MDX templates as `InstallationSelect`) wraps the refactored `Select` component, and allows the user to select the PouchDB document for an installation.

The `InstallationContext.Consumer`, `InstallationProvider` and `StoreProvider` components are exposed to MDX templates.

The other MDX templates are refactored to render the `CombustionSafetyChecksReport` component using the selected installation.